+++ lib/appydave/tools.rb
@@ -17,7 +17,6 @@ require 'appydave/tools/configuration/config'
 require 'appydave/tools/configuration/models/config_base'
 require 'appydave/tools/configuration/models/settings_config'
 require 'appydave/tools/configuration/models/bank_reconciliation_config'
-require 'appydave/tools/configuration/models/channel_projects_config'
 require 'appydave/tools/configuration/models/channels_config'
 require 'appydave/tools/name_manager/project_name'
 
@@ -25,8 +24,7 @@ Appydave::Tools::Configuration::Config.set_default do |config|
   config.config_path = File.expand_path('~/.config/appydave')
   config.register(:settings, Appydave::Tools::Configuration::Models::SettingsConfig)
   config.register(:bank_reconciliation, Appydave::Tools::Configuration::Models::BankReconciliationConfig)
-  config.register(:channels, Appydave::Tools::Configuration::ChannelsConfig)
-  config.register(:channel_projects, Appydave::Tools::Configuration::ChannelProjectsConfig)
+  config.register(:channels, Appydave::Tools::Configuration::Models::ChannelsConfig)
 end
 
 module Appydave