+++ spec/spec_helper.rb
@@ -13,7 +13,6 @@ Appydave::Tools::Configuration::Config.set_default do |config|
   config.register(:settings, Appydave::Tools::Configuration::Models::SettingsConfig)
   config.register(:bank_reconciliation, Appydave::Tools::Configuration::Models::BankReconciliationConfig)
   config.register(:channels, Appydave::Tools::Configuration::Models::ChannelsConfig)
-  config.register(:channel_projects, Appydave::Tools::Configuration::Models::ChannelProjectsConfig)
 end
 
 RSpec.configure do |config|